requirements.txt
pytest
selenium
webdriver-manager
eyes-selenium
flask
PyJWT
google-genai